@@ -1,3 +1,3 @@
 module github.com/Bios-Marcel/goclipimg
 
-go 1.12
+go 1.19
